VQA_enhancement/requirements.txt
numpy~=1.22.3
pandas~=1.4.2
matplotlib~=3.5.1
web.py~=0.62
tensorflow_hub
tensorflow-gpu
tensorflow-text
requests~=2.27.1
transformers~=4.18.0

deep-translator
torch
